Rename application id
