+++ app/src/main/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="org.mozilla.guardian">
+    package="org.mozilla.firefox.vpn">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="org.mozilla.guardian.MainApplication"
+        android:name=".GuardianApp"
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
