+++ app/src/main/java/org/mozilla/firefox/vpn/main/MainActivity.kt
@@ -1,0 +1,38 @@
+package org.mozilla.firefox.vpn.main
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import org.mozilla.firefox.vpn.R
+import org.mozilla.firefox.vpn.device.data.DeviceRepository
+import org.mozilla.firefox.vpn.device.domain.AddDeviceUseCase
+import org.mozilla.firefox.vpn.user.data.UserRepository
+
+class MainActivity : AppCompatActivity() {
+
+    private val addDevice: AddDeviceUseCase by lazy {
+        AddDeviceUseCase(
+            DeviceRepository(applicationContext),
+            UserRepository(applicationContext)
+        )
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val navView: BottomNavigationView = findViewById(R.id.nav_view)
+        val navController = findNavController(R.id.nav_host_fragment)
+        navView.setupWithNavController(navController)
+    }
+
+    companion object {
+        private const val TAG = "MainActivity"
+
+        fun getStartIntent(context: Context) = Intent(context, MainActivity::class.java)
+    }
+}
