+++ app/src/main/java/org/mozilla/firefox/vpn/main/settings/SettingsViewModel.kt
@@ -1,0 +1,13 @@
+package org.mozilla.firefox.vpn.main.settings
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SettingsViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is Settings Fragment"
+    }
+    val text: LiveData<String> = _text
+}
