--- app/src/main/java/org/mozilla/guardian/main/settings/SettingsViewModel.kt
@@ -1,13 +1,0 @@
-package org.mozilla.guardian.main.settings
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class SettingsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is Settings Fragment"
-    }
-    val text: LiveData<String> = _text
-}
