--- app/src/main/java/org/mozilla/guardian/main/vpn/VpnViewModel.kt
@@ -1,13 +1,0 @@
-package org.mozilla.guardian.main.vpn
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class VpnViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "Click to toggle VPN on/off"
-    }
-    val text: LiveData<String> = _text
-}
