--- app/src/main/java/org/mozilla/guardian/user/domain/GetLoginInfoUseCase.kt
@@ -1,13 +1,0 @@
-package org.mozilla.guardian.user.domain
-
-import org.mozilla.guardian.user.data.LoginInfo
-import org.mozilla.guardian.user.data.UserRepository
-
-class GetLoginInfoUseCase(
-    private val userRepository: UserRepository
-) {
-
-    suspend operator fun invoke(): LoginInfo {
-        return userRepository.getLoginInfo()
-    }
-}
